Todas son bases

+++ src/main/java/ar/alex/biblioteca/api/dto/CondicionesPrestamoDto.java
@@ -9,7 +9,7 @@ import lombok.Getter;
 @Builder
 @AllArgsConstructor
 public class CondicionesPrestamoDto {
-    private Long idCategoria;
+    private int idCategoria;
     private int numMaximoDiasPrestamo;
     private int numMaximoRenovaciones;
 }
